requirements.txt
Flask~=3.0.3
pillow~=10.3.0
WTForms~=3.1.2
matplotlib~=3.9.0
numpy~=1.26.4
Flask-WTF
Fask-Uploads